generator refactored to files, fixes #2204
