+++ src/stan/lang/generator/generate_model_typedef.hpp
@@ -2,6 +2,7 @@
 #define STAN_LANG_GENERATOR_GENERATE_MODEL_TYPEDEF_HPP
 
 #include <stan/lang/ast.hpp>
+#include <stan/lang/generator/constants.hpp>
 #include <ostream>
 #include <string>
 
@@ -18,7 +19,7 @@ namespace stan {
     void generate_model_typedef(const std::string& model_name,
                                 std::ostream& o) {
       o << "typedef " << model_name << "_namespace::" << model_name
-        << " stan_model;" <<EOL2;
+        << " stan_model;" << EOL2;
     }
 
   }
