+++ src/stan/lang/generator/init_local_var_visgen.hpp
@@ -5,6 +5,10 @@
 #include <stan/lang/generator/constants.hpp>
 #include <stan/lang/generator/generate_indent.hpp>
 #include <stan/lang/generator/generate_void_statement.hpp>
+#include <stan/lang/generator/has_lb.hpp>
+#include <stan/lang/generator/has_lub.hpp>
+#include <stan/lang/generator/has_ub.hpp>
+#include <stan/lang/generator/to_string.hpp>
 #include <stan/lang/generator/visgen.hpp>
 #include <ostream>
 #include <string>
