+++ src/stan/lang/generator/validate_transformed_params_visgen.hpp
@@ -4,6 +4,7 @@
 #include <stan/lang/ast.hpp>
 #include <stan/lang/generator/constants.hpp>
 #include <stan/lang/generator/generate_indent.hpp>
+#include <stan/lang/generator/visgen.hpp>
 #include <ostream>
 #include <string>
 #include <vector>
