+++ src/stan/lang/generator/write_array_visgen.hpp
@@ -4,6 +4,10 @@
 #include <stan/lang/ast.hpp>
 #include <stan/lang/generator/constants.hpp>
 #include <stan/lang/generator/generate_indent.hpp>
+#include <stan/lang/generator/has_lb.hpp>
+#include <stan/lang/generator/has_lub.hpp>
+#include <stan/lang/generator/has_ub.hpp>
+#include <stan/lang/generator/to_string.hpp>
 #include <stan/lang/generator/visgen.hpp>
 #include <ostream>
 #include <string>
